requirements.txt
azure-cognitiveservices-vision-computervision
azure-ai-translation-text==1.0.0b1
flask